more coverage

+++ porter.py
@@ -9,11 +9,15 @@ import os
 import shutil
 
 
-__all__ = ['mkdir', 'copy', 'copy_to', 'move', 'move_to', 'FileExistsError']
+__all__ = ['mkdir', 'copy', 'copy_to', 'move', 'move_to', 'FileExists', 'FileNotFound']
 __version__ = '0.0.1'
 
 
-class FileExistsError(EnvironmentError):
+class FileExists(EnvironmentError):
+    pass
+
+
+class FileNotFound(EnvironmentError):
     pass
 
 
@@ -27,7 +31,7 @@ def mkdir(directory, ignore=False, force=False):
             try:
                 os.mkdir(directory)
             except OSError:
-                raise FileExistsError("%s is exist." % directory)
+                raise FileExists("'%s' is exist." % directory)
         elif force:
             shutil.rmtree(directory)
             os.mkdir(directory)
@@ -42,10 +46,13 @@ def copy(src, dst, ignore=False, force=False):
         - copy a file to a future destination.
         - copy a directory to a future destination.
     """
+    if not os.path.exists(src) and not ignore:
+        raise FileNotFound("'%s' is not found." % src)
+
     if os.path.isfile(src):
         if os.path.exists(dst):
             if not ignore and not force:
-                raise FileExistsError("%s is exist." % dst)
+                raise FileExists("'%s' is exist." % dst)
             elif force:
                 shutil.copyfile(src, dst)
         else:
@@ -56,7 +63,7 @@ def copy(src, dst, ignore=False, force=False):
                 try:
                     shutil.copytree(src, dst)
                 except OSError:
-                    raise FileExistsError("%s is exist." % dst)
+                    raise FileExists("'%s' is exist." % dst)
             elif force:
                 shutil.rmtree(dst)
                 shutil.copytree(src, dst)
